fix: auction placebid service
